README.md
This is a console application that models a Grade Point Average (GPA) calculator and prints the result. Users are able to input course name and code, course unit, course score; one after the other and the calculator computes the score grade, records it and displays the cumulative GPA result after every successful input.